+++ go/service/ctl.go
@@ -4,10 +4,14 @@
 package service
 
 import (
-	"github.com/keybase/client/go/engine"
-	"github.com/keybase/client/go/libkb"
-	rpc "github.com/keybase/go-framed-msgpack-rpc"
 	"golang.org/x/net/context"
+
+	"github.com/keybase/client/go/engine"
+	"github.com/keybase/client/go/install"
+	"github.com/keybase/client/go/launchd"
+	"github.com/keybase/client/go/libkb"
+	keybase1 "github.com/keybase/client/go/protocol"
+	rpc "github.com/keybase/go-framed-msgpack-rpc"
 )
 
 type CtlHandler struct {
@@ -24,8 +28,32 @@ func NewCtlHandler(xp rpc.Transporter, v *Service, g *libkb.GlobalContext) *CtlH
 	}
 }
 
-// Stop is located in ctl_osx.go and ctl_non_osx.go for
-// OS-specific service stop.
+// Stop is called on the rpc keybase.1.ctl.stop, which shuts down the service.
+func (c *CtlHandler) Stop(_ context.Context, args keybase1.StopArg) error {
+	c.G().Log.Debug("Received stop(%d) RPC; shutting down", args.ExitCode)
+
+	if c.service.ForkType == keybase1.ForkType_LAUNCHD {
+		c.stopLaunchd()
+	}
+
+	go c.service.Stop(args.ExitCode)
+
+	return nil
+}
+
+func (c *CtlHandler) stopLaunchd() {
+	status := install.KeybaseServiceStatus(c.G(), c.G().Env.GetLabel())
+	if status.Pid == "" {
+		c.G().Log.Debug("Service does not appear to be running via launchd (label = %q)", c.G().Env.GetLabel())
+		return
+	}
+
+	c.G().Log.Debug("Removing %s from launchd", status.Label)
+	svc := launchd.NewService(status.Label)
+	if err := svc.Stop(false); err != nil {
+		c.G().Log.Warning("error stopping launchd service:", err)
+	}
+}
 
 func (c *CtlHandler) LogRotate(_ context.Context, sessionID int) error {
 	return c.G().Log.RotateLogFile()
